logs: only show each log's first line by default (expandable by JS)
When tracing is enabled in Yii, each log message takes several
lines, the first one being the real message, and the following ones the
trace.
This changeset hides the trace details by default. Clicking on a
line will toggle it locally. A JS link at the top of the column toggles
it globally.